--- __mocks__/openmrs-esm-api.mock.tsx
@@ -1,9 +1,0 @@
-import { of } from "rxjs";
-
-export function openmrsFetch() {
-  return new Promise(() => {});
-}
-
-export function getCurrentUser() {
-  return of({ authenticated: false });
-}
