--- __mocks__/openmrs-esm-error-handling.mock.tsx
@@ -1,3 +1,0 @@
-export function createErrorHandler() {
-  return true;
-}
