+++ __mocks__/openmrs-esm-framework.mock.tsx
@@ -1,4 +1,17 @@
 import React from "react";
+import { of } from "rxjs";
+
+export function openmrsFetch() {
+  return new Promise(() => {});
+}
+
+export function getCurrentUser() {
+  return of({ authenticated: false });
+}
+
+export function createErrorHandler() {
+  return true;
+}
 
 export const ComponentContext = React.createContext(null);
