+++ jest.config.json
@@ -4,13 +4,10 @@
   },
   "transformIgnorePatterns": [],
   "moduleNameMapper": {
+    "lodash-es": "lodash",
     "^@carbon/icons-react/es/(.*)$": "@carbon/icons-react/lib/$1",
     "^carbon-components-react/es/(.*)$": "carbon-components-react/lib/$1",
-    "@openmrs/esm-api": "<rootDir>/__mocks__/openmrs-esm-api.mock.tsx",
-    "@openmrs/esm-error-handling": "<rootDir>/__mocks__/openmrs-esm-error-handling.mock.tsx",
-    "@openmrs/esm-config": "<rootDir>/__mocks__/openmrs-esm-config.mock.tsx",
-    "@openmrs/esm-context": "<rootDir>/__mocks__/openmrs-esm-context.mock.tsx",
-    "@openmrs/esm-react-utils": "<rootDir>/__mocks__/openmrs-esm-react-utils.mock.tsx",
+    "@openmrs/esm-framework": "<rootDir>/__mocks__/openmrs-esm-framework.mock.tsx",
     "\\.(s?css)$": "identity-obj-proxy"
   },
   "setupFilesAfterEnv": ["<rootDir>/src/setupTests.ts"]
