+++ src/components/nav-header-panels/side-menu-panel.component.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SideNav from "carbon-components-react/es/components/UIShell";
+import { SideNav } from "carbon-components-react/es/components/UIShell";
 import { SideNavProps } from "carbon-components-react";
 import { ExtensionSlot } from "@openmrs/esm-framework";
 import styles from "./side-menu-panel.component.scss";
